requirements.txt
numpy
pandas
cvxpy
streamlit
matplotlib
seaborn
psycopg2-binary
apimoex
PyPortfolioOpt